decouple sdl2 context creation and application
