+++ src/application.cpp
@@ -5,39 +5,15 @@
 #include <functional>
 #include "context.h"
 #include "default_font.h"
-#include "logger.h"
 #include "loop.h"
 #include "sdl_texture.h"
 #include "video_memory.h"
+#include "sdl_context.h"
 
 Application::Application(ApplicationListener* listener, const ApplicationConfig& config)
     : listener(std::unique_ptr<ApplicationListener>(listener)) {
-  if (SDL_Init(SDL_INIT_VIDEO) != 0) {
-    logSDLError("SDL_Init");
-    return;  // TODO: what? throw exception?
-  }
-  Uint32 flags = SDL_WINDOW_SHOWN;
-  if (config.is_fullscreen) {
-    flags |= SDL_WINDOW_FULLSCREEN_DESKTOP;
-  } else {
-    flags |= SDL_WINDOW_RESIZABLE;
-  }
-  win = std::unique_ptr<SDL_Window, sdl_deleter>(
-      SDL_CreateWindow(config.name.c_str(), SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED, config.GetWindowWidth(),
-                       config.GetWindowHeight(), flags),
-      sdl_deleter());
-  if (!win) {
-    logSDLError("SDL_CreateWindow");
-    return;  // TODO: what? throw exception?
-  }
-  ren = std::unique_ptr<SDL_Renderer, sdl_deleter>(
-      SDL_CreateRenderer(win.get(), -1, SDL_RENDERER_ACCELERATED | SDL_RENDERER_PRESENTVSYNC), sdl_deleter());
-  if (!ren) {
-    logSDLError("SDL_CreateRenderer");
-    return;  // TODO: what? throw exception?
-  }
   vram = std::unique_ptr<VideoMemory>(new VideoMemory(config.width, config.height));
-  texture = std::unique_ptr<SDLTexture>(new SDLTexture(ren, *vram));
+  sdl_context = std::unique_ptr<SDLContext>(new SDLContext(config, vram.get()));
 }
 Application::~Application() {
   SDL_Quit();
@@ -57,13 +33,8 @@ void Application::Update(Context& ctx, double delta_time) {
   if (!ctx.IsPaused()) {
     ctx.Tick(delta_time);
     listener->OnRender(ctx);
-    Render();
+    sdl_context->Render();
   }
-}
-void Application::Render() {
-  SDL_RenderClear(ren.get());
-  texture->Render();
-  SDL_RenderPresent(ren.get());
 }
 void Application::HandleEvents(Context& ctx) {
   SDL_Event e;
